+++ myapps/hello-next-js/global/featureFlags.ts
@@ -1,0 +1,4 @@
+export const TABLE_FILTER_OPTIMISATION = {
+    withUseDeferredValue: true, // recommended to optimise your filter feature 
+    withUseTransition: false, // not recommended for this use case scenario
+}
