+++ myapps/hello-next-js/pages/api/tasks/v1/bff/delete-rows.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { Task } from "@/app/types/Task";
-import { BASE_URL } from "../../../../../feature-flags/tasksBff";
+import { BASE_URL } from "../../../../../global/common";
 import { TASKS_BFF_HEADER } from "../../../../../global/common";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
