+++ myapps/hello-next-js/pages/api/tasks/v1/bff/index.ts
@@ -6,7 +6,7 @@
 // /pages/api/tasks/v1/bff/         ← safe route for frontend to hit (calls tasks/v1/sql)
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { Task } from "@/app/types/Task";
-import { BASE_URL } from "../../../../../feature-flags/tasksBff";
+import { BASE_URL } from "../../../../../global/common";
 import { TASKS_BFF_HEADER } from "../../../../../global/common";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
