+++ myapps/hello-next-js/pages/task-crud-fullstack/with-swr.tsx
@@ -4,6 +4,7 @@ import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
 import { SWRConfig } from 'swr';
 import { TaskModel } from '@/app/models/Task/use-client/TaskModel';
 import { TaskWithSWRPage } from '@/app/views/Task/use-client/taskWithSWRPage';
+import { BASE_URL } from "../../global/common";
 
 // dev note: alternative function signature:
 //type Props = {
@@ -22,7 +23,7 @@ const TaskListWithSWR = ({ fallback }: InferGetServerSidePropsType<typeof getSer
 };
 
 export const getServerSideProps: GetServerSideProps = async () => {
-  const tasks = await (new TaskModel()).getTasksDBRows();
+  const tasks = await (new TaskModel()).getTasksDBRows(`${BASE_URL}/api/tasks/v1/bff`);
   
   return {
     props: {
