+++ myapps/hello-next-js/src/app/models/Task/use-client/TaskModel.test.ts
@@ -1,5 +1,5 @@
 import { TaskModel } from './TaskModel';
-import { DATA_FETCH_MODE, BASE_URL } from "../../../../../feature-flags/tasksBff";
+import { BASE_URL } from "../../../../../global/common";
 
 // Define mock response type
 type MockResponse = {
@@ -13,7 +13,7 @@ describe('TaskModel', () => {
   let taskModel: TaskModel;
   let spyConsoleError: jest.SpyInstance<any, any>;
 
-  const url = DATA_FETCH_MODE === "getServerSideProps" ? BASE_URL : "";
+  const url = process.env.NODE_ENV === "production" ? BASE_URL : "";
   
   // Helper function to create mock responses
   const mockResponse = (ok: boolean, data: any): MockResponse => ({
