+++ myapps/hello-next-js/src/app/views/Task/use-client/taskDetailPage.tsx
@@ -1,5 +1,4 @@
 'use client';
-// The View connects the ViewModel and UI component
 import { useState, useEffect } from 'react';
 import Link from 'next/link';
 //import { useTaskViewModel } from '@/app/viewModels/Task/use-client/useTasksViewModel';
