+++ myapps/hello-next-js/src/app/views/Task/use-client/taskPage.tsx
@@ -1,5 +1,4 @@
 'use client';
-// The View connects the ViewModel and UI component
 import { useState, useEffect } from "react";
 import { useTaskViewModel } from '@/app/viewModels/Task/use-client/useTasksViewModel';
 import { TaskSeedDB } from '@/app/components/TaskSeedDB';
